Pipeline

On this page, we explain how to analyze a given CY5 space.

Mathematical setup

We recall the mathematical setup of [HS26], to which we refer for full details. Let $X$ be a smooth equivariantly Calabi-Yau GKM space with respect to the torus $T$, and let $\epsilon_1,\dots,\epsilon_r$ be the equivariant parameters for $T$. Given a curve class $\beta$ and genus $g\ge 0$, let

\[GW_{g,\beta}(X,T) \in \mathbb{Q}(\epsilon_1,\dots,\epsilon_r)\]

be the associated $T$-equivariant Gromov-Witten invariant (with connected domain curves) for $X$ in genus $g$ and curve class $\beta$. Let $u$ and $Q$ be formal variables. Then

\[\Omega_\beta(\epsilon, u)\in \mathbb{Q}(\epsilon_1,\dots,\epsilon_r)[\![u]\!]\]

is defined by

\[\sum_{\beta\neq 0}\sum_{g\ge 0} Q^\beta u^{2g-2} GW_{g,\beta}(X,T) = \sum_{\beta\neq 0} \sum_{k>0} \frac{1}{k} Q^{k\beta} \Omega_\beta(\epsilon,ku)\]

The point of this package is to test conjectural formulae for $\Omega_\beta$ in specific examples.

Note

The conjectural formulae for $\Omega_\beta$ are formulated in terms of the substitution $q_i=e^{u\epsilon_i}$. As GKMtools.jl computes $GW_{g,\beta}(X,T)$ using the equivariant parameters $\epsilon_i$, we only perform this substitution at the end when we compare the conjectural formula to the computed low-order terms of $\Omega_\beta$ in $u$.

Implementation

Our pipeline has the following structure.

Define the $T$-space $X$ using GKMtools.jl.

Define the conjectured values of $\Omega_\beta$.

Compute $GW_{g,\beta}(X,T)$ using GKMtools.jl for small values of $g,\beta$.

Solve the above recursion to compute $\Omega_\beta$ up to some order of $u$.

Compare the results.

Steps 1. and 2. are implemented as one function for each family of spaces that we check in the paper (see Spaces). Steps 3.-5. are implemented in the function get_Omega_beta below.

The main function

The following function is the main function used in our numerical experiments to confirm conjectural values of $\Omega_\beta$ up to a certain genus.

GW_CY5.get_Omega_beta — Function

get_Omega_beta(G::AbstractGKM_graph, betas::Vector, gMax::Int64; check_predictions::Bool=false)

Return a dictionary assigning to each curve class $\beta$ in betas the invariant $\Omega_\beta$ as rational function in the equivariant parameters and the genus parameter $u$. Internally, this function uses get_GW_beta to get the Gromov-Witten invariants required to compute $\Omega_\beta$.

The output also contains $\Omega_\beta$ for every $\beta$ for which some positive integer multiple is contained in beta.

Arguments

G: the GKM graph of the space $X$ whose $\Omega_\beta$ should be computed.

betas: the vector of curve classes on $X$ for which $\Omega_\beta$ should be computed.

gMax: the maximum genus up to which each $\Omega_\beta$ will be approximated. That is, the highest order term will be $u^{2\text{gMax}-2}$.

check_predictions: this optional argument with default value false controls whether the computed values of $\Omega_\beta$ are compared to the conjectured values. If check_predictions=true, the conjectured values are retrieved from get_Omega_prediction.

show_bar: this optional argument can be set to false to disable the progress bar of the Gromov-Witten localization computation.

Example 1: Computing $\Omega_\beta$

Let us compute the genus zero and genus one terms of $\Omega_\beta$ for the example $Z=\mathcal{A}_2\times \mathbb{C}^3$ (see Spaces), for some examples of $\beta$.

julia> G = Ar_times_C3(2);

julia> b = curve_class(G, Edge(1, 2))
(1, 0)

julia> c = curve_class(G, Edge(2, 3))
(0, 1)

julia> max_genus = 1;

julia> Omega = get_Omega_beta(G, [3*b, 4*b, b+c], max_genus; show_bar=false);
Calculating b=(3, 0), g=0
Calculating b=(3, 0), g=1
Calculating b=(1, 0), g=0
Calculating b=(1, 0), g=1
Calculating b=(4, 0), g=0
Calculating b=(4, 0), g=1
Calculating b=(2, 0), g=0
Calculating b=(2, 0), g=1
Calculating b=(1, 1), g=0
Calculating b=(1, 1), g=1

julia> Omega[b]
(1//12*t2^2*t4*u^2 + 1//12*t2^2*t5*u^2 + 1//12*t2*t4^2*u^2 + 1//12*t2*t4*t5*u^2 + 1//12*t2*t5^2*u^2 - t2)//(t2*t4*t5*u^2 + t4^2*t5*u^2 + t4*t5^2*u^2)

julia> Omega[2*b]
0

julia> Omega[3*b]
0

julia> Omega[4*b]
0

julia> Omega[b+c]
(1//12*t2^2*t4*u^2 + 1//12*t2^2*t5*u^2 + 1//12*t2*t4^2*u^2 + 1//12*t2*t4*t5*u^2 + 1//12*t2*t5^2*u^2 - t2)//(t2*t4*t5*u^2 + t4^2*t5*u^2 + t4*t5^2*u^2)

The equivariant parameters $\epsilon_1,\epsilon_2,\dots$ are displayed as t1, t2, etc.

Example 2: Checking conjectured values of $\Omega_\beta$

Setting check_predictios=true, we can automatically check if $\Omega_\beta$ matches the output of get_Omega_prediction:

julia> get_Omega_beta(G, [3*b, 4*b, b+c], max_genus; show_bar=false, check_predictions=true);
Calculating b=(3, 0), g=0
Calculating b=(3, 0), g=1
Calculating b=(1, 0), g=0
Calculating b=(1, 0), g=1
Calculating b=(4, 0), g=0
Calculating b=(4, 0), g=1
Calculating b=(2, 0), g=0
Calculating b=(2, 0), g=1
Calculating b=(1, 1), g=0
Calculating b=(1, 1), g=1
Prediction holds for (4, 0)
Prediction holds for (2, 0)
Prediction holds for (3, 0)
Prediction holds for (1, 1)
Prediction holds for (1, 0)
All predictions hold.

In this case, all predicticted values of $\Omega_\beta$ hold in the tested range.

Example 3: Behavior when the conjecture fails

Let us also see an example in which the conjectured value of $\Omega_\beta$ does not agree with the true value computed from Gromov-Witten invariants. We take

\[Z=\text{Tot}_{\mathbb{P}^1\times\mathbb{P}^1}(\mathcal{O}(-2,0)\oplus\mathcal{O}(0,-2)\oplus\mathcal{O})\]

and let $\beta$ be the curve class of one of the two factor $\mathbb{P}^1$s. Then the properness assumption of [HS26, Conjecture 5.1] does not hold for $\beta$ and its multiples. Let us denote $\beta$ by b in the following code.

julia> P1_P1 = hirzebruch_surface(NormalToricVariety, 0)
Normal toric variety

julia> P = picard_group(P1_P1)
Z^2

julia> L1 = toric_line_bundle(P1_P1, -2*P[1]) # O(-2, 0)
Toric line bundle on a normal toric variety

julia> L2 = toric_line_bundle(P1_P1, -2*P[2]) # O(0, -2)
Toric line bundle on a normal toric variety

julia> CY4 = total_space(gkm_vector_bundle_of_toric([L1, L2]))
GKM graph with 4 nodes, valency 4 and axial function:
2 -> 1 => (1, 0, 0, 0, -1, 0)
3 -> 2 => (0, 1, 0, 0, 0, -1)
4 -> 1 => (0, 1, 0, 0, 0, -1)
4 -> 3 => (-1, 0, 0, 0, 1, 0)
Standalone flags:
1.3 => (0, 0, -1, 0, -2, 0)
1.4 => (0, 0, 0, -1, 0, -2)
2.3 => (-2, 0, -1, 0, 0, 0)
2.4 => (0, 0, 0, -1, 0, -2)
3.3 => (-2, 0, -1, 0, 0, 0)
3.4 => (0, -2, 0, -1, 0, 0)
4.3 => (0, 0, -1, 0, -2, 0)
4.4 => (0, -2, 0, -1, 0, 0)

julia> G = CY5_from_CY4(CY4; equiCY=true)
GKM graph with 4 nodes, valency 5 and axial function:
2,1 -> 1,1 => (1, 0, 0, 0, -1, 0, 0)
3,1 -> 2,1 => (0, 1, 0, 0, 0, -1, 0)
4,1 -> 1,1 => (0, 1, 0, 0, 0, -1, 0)
4,1 -> 3,1 => (-1, 0, 0, 0, 1, 0, 0)
Standalone flags:
1,1.3 => (0, 0, -1, 0, -2, 0, 0)
1,1.4 => (0, 0, 0, -1, 0, -2, 0)
1,1.5 => (1, 1, 1, 1, 1, 1, 0)
2,1.3 => (-2, 0, -1, 0, 0, 0, 0)
2,1.4 => (0, 0, 0, -1, 0, -2, 0)
2,1.5 => (1, 1, 1, 1, 1, 1, 0)
3,1.3 => (-2, 0, -1, 0, 0, 0, 0)
3,1.4 => (0, -2, 0, -1, 0, 0, 0)
3,1.5 => (1, 1, 1, 1, 1, 1, 0)
4,1.3 => (0, 0, -1, 0, -2, 0, 0)
4,1.4 => (0, -2, 0, -1, 0, 0, 0)
4,1.5 => (1, 1, 1, 1, 1, 1, 0)

julia> b = curve_class(G, Edge(1, 2))
(0, 1)

julia> Omega = get_Omega_beta(G, [b, 2*b, 3*b], 1; check_predictions = true, show_bar=false);
Calculating b=(0, 1), g=0
Calculating b=(0, 1), g=1
Calculating b=(0, 2), g=0
Calculating b=(0, 2), g=1
Calculating b=(0, 3), g=0
Calculating b=(0, 3), g=1
Prediction holds for (0, 2)
Prediction holds for (0, 3)
Prediction fails for (0, 1):
  prediction = 0

julia> Omega[b]
(1//12*t1^2*t2*u^2 + 1//12*t1^2*t4*u^2 + 1//12*t1^2*t6*u^2 + 1//12*t1*t2^2*u^2 + 1//6*t1*t2*t3*u^2 + 1//6*t1*t2*t5*u^2 - 1//6*t1*t2*t6*u^2 + 1//6*t1*t3*t4*u^2 + 1//6*t1*t3*t6*u^2 + 1//6*t1*t4*t5*u^2 + 1//6*t1*t5*t6*u^2 + 1//12*t1*t6^2*u^2 - t1 + 1//12*t2^2*t3*u^2 + 1//12*t2^2*t5*u^2 + 1//12*t2*t3^2*u^2 + 1//6*t2*t3*t5*u^2 - 1//6*t2*t3*t6*u^2 + 1//12*t2*t5^2*u^2 - 1//6*t2*t5*t6*u^2 + 1//12*t3^2*t4*u^2 + 1//12*t3^2*t6*u^2 + 1//6*t3*t4*t5*u^2 + 1//6*t3*t5*t6*u^2 + 1//12*t3*t6^2*u^2 - t3 + 1//12*t4*t5^2*u^2 + 1//12*t5^2*t6*u^2 + 1//12*t5*t6^2*u^2 - t5)//(t1*t2*t4*u^2 + 2*t1*t2*t6*u^2 + 1//2*t1*t4^2*u^2 + t1*t4*t6*u^2 + t2^2*t4*u^2 + 2*t2^2*t6*u^2 + t2*t3*t4*u^2 + 2*t2*t3*t6*u^2 + 3//2*t2*t4^2*u^2 + t2*t4*t5*u^2 + 4*t2*t4*t6*u^2 + 2*t2*t5*t6*u^2 + 2*t2*t6^2*u^2 + 1//2*t3*t4^2*u^2 + t3*t4*t6*u^2 + 1//2*t4^3*u^2 + 1//2*t4^2*t5*u^2 + 3//2*t4^2*t6*u^2 + t4*t5*t6*u^2 + t4*t6^2*u^2)

julia> Omega[2*b]
0

julia> Omega[3*b]
0

As we see, the conjectured value for $\Omega_\beta$ is zero, but the true value is non-zero. The printed output highlights whenever the conjectured value is incorrect.

Mathematically, this example is not a counterexample to [HS26, Conjecture 5.1] because one of the conjecture's assumptions is not satisfied in this case.

Most important helper functions

The two most important helper functions used by get_Omega_beta are:

get_GW_beta: uses GKMtools.jl to compute individual Gromov-Witten invariants and packages them in a truncated generating series.

get_Omega_prediction: provides the conjectured values of $\Omega_\beta$ for comparison to the actual values computed from Gromov-Witten invariants.

GW_CY5.get_GW_beta — Function

get_GW_beta(G::AbstractGKM_graph, betas::Vector{T}, gMax::Int64; show_bar::Bool=true) where T <: CC

Let $(Z,T)$ be a GKM space with GKM graph $G$. Return the truncated Gromov–Witten series

\[\sum_{g \ge 0}^{\text{gMax}}u^{2g-2} GW_{g,\beta}(Z,T)\]

for each curve class $\beta$ listed in beta. The individual Gromov-Witten invariants are computed using the function gromov_witten of GKMtools.

The output also contains the truncated Gromov-Witten series for each curve class with some positive integer multiple contained in betas.

Arguments

G: the GKM graph of the GKM space $(Z,T)$ whose $GW_\beta(Z,T)$ should be computed.

betas: the vector of curve classes on $Z$ for which $GW_\beta(Z,T)$ should be computed.

gMax: the maximum genus up to which each $GW_\beta(Z,T)$ will be approximated. That is, the highest order term will be $u^{2\text{gMax}-2}$.

show_bar: this optional argument can be set to false to disable the progress bar of the Gromov-Witten localization computation.

Example

Let us compute the Gromov-Witten series in degrees $\beta=1,2,3$ for the example

\[Z = \text{Tot}_{\mathbb{P}^1}(\mathcal{O}(-2))\times\mathbb{C}^3\]

and the $T$-action defined in [HS26, Example 1.1]. Its GKM graph can be obtained using gkm_graph_of_example_1_1.

We compute the Gromov-Witten series up to genus 1. That is, the powers of $u$ that appear are $u^{-2}$ and $1$.

julia> G = gkm_graph_of_example_1_1();

julia> beta = curve_class(G, Edge(1, 2));

julia> gMax = 1;

julia> GW = get_GW_beta(G, [beta, 2*beta, 3*beta], gMax; show_bar=false);
Calculating b=(1), g=0
Calculating b=(1), g=1
Calculating b=(2), g=0
Calculating b=(2), g=1
Calculating b=(3), g=0
Calculating b=(3), g=1

julia> GW[beta]
(1//12*t1*t4*t5*u^2 + 1//12*t1*t4*t6*u^2 + 1//12*t1*t5*t6*u^2 + t1 + 1//12*t2*t4*t5*u^2 + 1//12*t2*t4*t6*u^2 + 1//12*t2*t5*t6*u^2 + t2 + 1//12*t3*t4*t5*u^2 + 1//12*t3*t4*t6*u^2 + 1//12*t3*t5*t6*u^2 + t3)//(t4*t5*t6*u^2)

julia> GW[2*beta]
(1//24*t1*t4*t5*u^2 + 1//24*t1*t4*t6*u^2 + 1//24*t1*t5*t6*u^2 + 1//8*t1 + 1//24*t2*t4*t5*u^2 + 1//24*t2*t4*t6*u^2 + 1//24*t2*t5*t6*u^2 + 1//8*t2 + 1//24*t3*t4*t5*u^2 + 1//24*t3*t4*t6*u^2 + 1//24*t3*t5*t6*u^2 + 1//8*t3)//(t4*t5*t6*u^2)

julia> GW[3*beta]
(1//36*t1*t4*t5*u^2 + 1//36*t1*t4*t6*u^2 + 1//36*t1*t5*t6*u^2 + 1//27*t1 + 1//36*t2*t4*t5*u^2 + 1//36*t2*t4*t6*u^2 + 1//36*t2*t5*t6*u^2 + 1//27*t2 + 1//36*t3*t4*t5*u^2 + 1//36*t3*t4*t6*u^2 + 1//36*t3*t5*t6*u^2 + 1//27*t3)//(t4*t5*t6*u^2)

GW_CY5.get_Omega_prediction — Function

get_Omega_prediction(G::AbstractGKM_graph, t::Vector, u, beta, max_genus::Int64)

Return the conjectural prediction of $\Omega_\beta$ for the space with GKM graph G.

Arguments

G: the GKM graph of the space $X$ whose $\Omega_\beta$ we care about.

t: the vector of equivariant parameters.

u: the formal variable $u$ that keeps track of the genus.

beta: the curve class $\beta$ for which the conjectural value of $\Omega_\beta$ should be returned.

max_genus: the maximum genus up to which the conjectural value of $\Omega_\beta$ should be returned. That is, the highest exponent of $u$ in the returned value is 2*max_genus - 1.

Note

This function requires the attribute :example_type of G to be set. This holds automatically whenever G is obtained from one of the constructors in the list of implemented spaces.

Example

Let us get the predicted $\Omega_\beta$ for $Z=\mathcal{A}_2\times \mathbb{C}^3$, continuing the example from get_Omega_beta.

julia> G = Ar_times_C3(2);

julia> b = curve_class(G, Edge(1, 2));

julia> c = curve_class(G, Edge(2, 3));

julia> max_genus = 1;

julia> R, (t1, t2, t3, t4, t5, u) = polynomial_ring(QQ, ["t1", "t2", "t3", "t4", "t5", "u"])
(Multivariate polynomial ring in 6 variables over QQ, QQMPolyRingElem[t1, t2, t3, t4, t5, u])

julia> prediction = get_Omega_prediction(G, [t1, t2, t3, t4, t5], u, b, max_genus)
(-1//12*t3^2*t4*u^2 - 1//12*t3^2*t5*u^2 - 1//12*t3*t4^2*u^2 - 1//4*t3*t4*t5*u^2 - 1//12*t3*t5^2*u^2 - t3 - 1//12*t4^2*t5*u^2 - 1//12*t4*t5^2*u^2 - t4 - t5)//(t3*t4*t5*u^2)

This does not yet match the result of get_Omega_beta above, because we still need to make the equivariantly Calabi-Yau substitution. The appropriate substitution is stored in G:

julia> substitution = get_attribute(G, :equiCY_substitution)
5-element Vector{QQMPolyRingElem}:
 t1
 t2
 -t2 - t4 - t5
 t4
 t5

julia> # Translate the substitution vector to our own variables.
        substitution = [evaluate(t, [t1, t2, t3, t4, t5]) for t in substitution]
5-element Vector{QQMPolyRingElem}:
 t1
 t2
 -t2 - t4 - t5
 t4
 t5

julia> evaluate(prediction, vcat(substitution, [u]))
(1//12*t2^2*t4*u^2 + 1//12*t2^2*t5*u^2 + 1//12*t2*t4^2*u^2 + 1//12*t2*t4*t5*u^2 + 1//12*t2*t5^2*u^2 - t2)//(t2*t4*t5*u^2 + t4^2*t5*u^2 + t4*t5^2*u^2)

This matches the result of the example in get_Omega_beta, as required.

The above is an illustration of how the pipeline works. When get_Omega_beta is used with check_predictions=true, it uses the same mechanism to determine if the actual $\Omega_\beta$ matches the conjectured one.

More details

A more detailed example of the pipeline in action can be found on the page Details of the Pipeline.