Spaces

On this page, we list the implemented spaces and constructor functions that are compatible with the Pipeline. This means that they can be used in conjunction with get_Omega_prediction and hence with get_Omega_beta and its optional argument check_predictions=true.

Note

Only those functions that return a 5-dimensional result are compatible with the Pipeline. Functions returning a lower-dimensional result are listed because they are ingredients for 5-dimensional constructions.

Strip geometries

GW_CY5.Ar_times_C1 — Function

Ar_times_C1(r::Int64) -> GKMtools.AbstractGKM_graph

Return $\mathcal{A}_r\times\mathbb{C}^1$, where $r$ is the number of rational curves in the chain.

Example

julia> G = Ar_times_C1(2)
GKM graph with 3 nodes, valency 3 and axial function:
2 -> 1 => (1, -1, 0)
3 -> 2 => (1, -2, 0)
Standalone flags:
1.2 => (0, 0, 1)
1.3 => (1, 0, 0)
2.3 => (0, 0, 1)
3.2 => (0, 0, 1)
3.3 => (-1, 3, 0)

GW_CY5.Ar_times_C3 — Function

Ar_times_C3(r::Int64) -> GKMtools.AbstractGKM_graph

Return $\mathcal{A}_r\times\mathbb{C}^3$, where $r$ is the number of rational curves in the chain. The space is equivariantly Calabi-Yau.

Example

julia> G = Ar_times_C3(2)
GKM graph with 3 nodes, valency 5 and axial function:
2,1 -> 1,1 => (1, -1, 0, 0, 0)
3,1 -> 2,1 => (1, -2, 0, 0, 0)
Standalone flags:
1,1.2 => (0, -1, 0, -1, -1)
1,1.3 => (1, 0, 0, 0, 0)
1,1.4 => (0, 0, 0, 1, 0)
1,1.5 => (0, 0, 0, 0, 1)
2,1.3 => (0, -1, 0, -1, -1)
2,1.4 => (0, 0, 0, 1, 0)
2,1.5 => (0, 0, 0, 0, 1)
3,1.2 => (0, -1, 0, -1, -1)
3,1.3 => (-1, 3, 0, 0, 0)
3,1.4 => (0, 0, 0, 1, 0)
3,1.5 => (0, 0, 0, 0, 1)

GW_CY5.gkm_3d_strip — Function

gkm_3d_strip(ups_and_downs::Vector{Int64}; equiCY::Bool=false) -> GKMtools.AbstractGKM_graph

Create the GKM graph of the 3d strip geometry described by ups_and_downs (see [HS26, Section 2.1]).

Arguments

ups_and_downs: a vector of non-zero integers with the following meaning in terms of [HS26, (18)]: there will be ups_and_downs[1] flags up (if positive) or down (if negative), followed by ups_and_downs[2] flags up or down, and so on.

equiCY: If this optional argument with default value false is set to true, the returned space will be linearized to be equivariantly Calabi-Yau.

Weight convention:

In terms of [HS26, (18)]: flags pointing up have weight $\epsilon_2$, flags pointing down have weight $\epsilon_3$, and the flag at vertex 1 pointing left has weight $\epsilon_1$.

Example

Let us see the same example once with equiCY=false (default) and once with equiCY=true.

julia> G = gkm_3d_strip([1, -1, 1]) # equiCY=false
GKM graph with 3 nodes, valency 3 and axial function:
2 -> 1 => (1, 0, -1)
3 -> 2 => (1, -1, -1)
Standalone flags:
1.2 => (0, 1, 0)
1.3 => (1, 0, 0)
2.3 => (0, 0, 1)
3.2 => (-1, 1, 2)
3.3 => (0, 1, 0)

julia> G = gkm_3d_strip([1, -1, 1]; equiCY=true)
GKM graph with 3 nodes, valency 3 and axial function:
2 -> 1 => (-1, 1)
3 -> 2 => (-1, 0)
Standalone flags:
1.2 => (0, 1)
1.3 => (-1, 0)
2.3 => (0, -1)
3.2 => (1, -1)
3.3 => (0, 1)

GW_CY5.gkm_5d_strip — Function

gkm_5d_strip(ups_and_downs::Vector{Int64}; equiCY::Bool=false) -> GKMtools.AbstractGKM_graph

Return the product of the output of gkm_3d_strip with $\mathbb{C}^2$, where the latter has weights $\epsilon_4, \epsilon_5$.

Arguments

ups_and_downs: Same meaning as in gkm_3d_strip

equiCY: If this optional argument with default value false is set to true, the result is linearized to be equivariantly Calabi-Yau. Note that the underlying 3d strip geometry is not taken to be equivariantly Calabi-Yau.

Example

Let us produce the example of gkm_3d_strip times $\mathbb{C}^2$.

julia> G = gkm_5d_strip([1, -1, 1])
GKM graph with 3 nodes, valency 5 and axial function:
2,1 -> 1,1 => (1, 0, -1, 0, 0)
3,1 -> 2,1 => (1, -1, -1, 0, 0)
Standalone flags:
1,1.2 => (0, 1, 0, 0, 0)
1,1.3 => (1, 0, 0, 0, 0)
1,1.4 => (0, 0, 0, 1, 0)
1,1.5 => (0, 0, 0, 0, 1)
2,1.3 => (0, 0, 1, 0, 0)
2,1.4 => (0, 0, 0, 1, 0)
2,1.5 => (0, 0, 0, 0, 1)
3,1.2 => (-1, 1, 2, 0, 0)
3,1.3 => (0, 1, 0, 0, 0)
3,1.4 => (0, 0, 0, 1, 0)
3,1.5 => (0, 0, 0, 0, 1)

GW_CY5.minus_one_minus_one_chain_3d — Function

minus_one_minus_one_chain_3d(r::Int64) -> GKMtools.AbstractGKM_graph

Return a chain of $\mathcal{O}(-1)\oplus\mathcal{O}(-1)$ bundles of length $r$. The result is Calabi-Yau but not equivariantly Calabi–Yau and coincides with the output of gkm_3d_strip with ups_and_downs = [-1, 1, -1, ...].

Example

Let us produce the chain of length 2.

julia> GW_CY5.minus_one_minus_one_chain_3d(2)
GKM graph with 3 nodes, valency 3 and axial function:
2 -> 1 => (1, -1, 0)
3 -> 2 => (1, -1, -1)
Standalone flags:
1.2 => (0, 0, 1)
1.3 => (1, 0, 0)
2.3 => (0, 1, 0)
3.2 => (-1, 2, 1)
3.3 => (0, 0, 1)

GW_CY5.minus_one_minus_one_chain — Function

minus_one_minus_one_chain(r::Int64) -> GKMtools.AbstractGKM_graph

Return a chain of $\mathcal{O}(-1)\oplus\mathcal{O}(-1)$ bundles of length $r$, product with $\mathbb{C}^2$. The result is equivariantly Calabi-Yau and coincides with the output of gkm_5d_strip with ups_and_downs = [-1, 1, -1, ...].

Example

Let us produce the chain of length 2.

julia> G = minus_one_minus_one_chain(2)
GKM graph with 3 nodes, valency 5 and axial function:
2,1 -> 1,1 => (1, -1, 0, 0, 0)
3,1 -> 2,1 => (1, 0, 0, 1, 1)
Standalone flags:
1,1.2 => (0, -1, 0, -1, -1)
1,1.3 => (1, 0, 0, 0, 0)
1,1.4 => (0, 0, 0, 1, 0)
1,1.5 => (0, 0, 0, 0, 1)
2,1.3 => (0, 1, 0, 0, 0)
2,1.4 => (0, 0, 0, 1, 0)
2,1.5 => (0, 0, 0, 0, 1)
3,1.2 => (-1, 1, 0, -1, -1)
3,1.3 => (0, -1, 0, -1, -1)
3,1.4 => (0, 0, 0, 1, 0)
3,1.5 => (0, 0, 0, 0, 1)

Closed vertex

GW_CY5.gkm_3d_closed_vertex — Function

gkm_3d_closed_vertex() -> GKMtools.AbstractGKM_graph

Return the GKM graph of the 3d closed vertex geometry. The weights are defined to agree with [HS26, Figure 4].

Example

julia> gkm_3d_closed_vertex()
GKM graph with 4 nodes, valency 3 and axial function:
0 -> 1 => (-1//2, 1//2, 1//2)
2 -> 0 => (-1//2, 1//2, -1//2)
3 -> 0 => (-1//2, -1//2, 1//2)
Standalone flags:
1.2 => (0, 1, 0)
1.3 => (0, 0, 1)
2.2 => (1, 0, 0)
2.3 => (0, 0, 1)
3.2 => (1, 0, 0)
3.3 => (0, 1, 0)

GW_CY5.gkm_5d_closed_vertex — Function

gkm_5d_closed_vertex(; equiCY::Bool=false) -> GKMtools.AbstractGKM_graph

Return the product of the output of gkm_3d_closed_vertex with $\mathbb{C}^2$, where the latter has weights $\epsilon_4, \epsilon_5$.

Arguments

equiCY: If this optional argument with default value false is set to true, the result is linearized to be equivariantly Calabi-Yau. Note that the underlying 3d closed vertex geometry is not taken to be equivariantly Calabi-Yau.

Example

Let us see the output of this function once with equiCY=false (default) and once with equiCY=true.

julia> gkm_5d_closed_vertex()
GKM graph with 4 nodes, valency 5 and axial function:
0,1 -> 1,1 => (-1//2, 1//2, 1//2, 0, 0)
2,1 -> 0,1 => (-1//2, 1//2, -1//2, 0, 0)
3,1 -> 0,1 => (-1//2, -1//2, 1//2, 0, 0)
Standalone flags:
1,1.2 => (0, 1, 0, 0, 0)
1,1.3 => (0, 0, 1, 0, 0)
1,1.4 => (0, 0, 0, 1, 0)
1,1.5 => (0, 0, 0, 0, 1)
0,1.4 => (0, 0, 0, 1, 0)
0,1.5 => (0, 0, 0, 0, 1)
2,1.2 => (1, 0, 0, 0, 0)
2,1.3 => (0, 0, 1, 0, 0)
2,1.4 => (0, 0, 0, 1, 0)
2,1.5 => (0, 0, 0, 0, 1)
3,1.2 => (1, 0, 0, 0, 0)
3,1.3 => (0, 1, 0, 0, 0)
3,1.4 => (0, 0, 0, 1, 0)
3,1.5 => (0, 0, 0, 0, 1)

julia> gkm_5d_closed_vertex(equiCY=true)
GKM graph with 4 nodes, valency 5 and axial function:
0,1 -> 1,1 => (-1, 0, 0, -1, -1)
2,1 -> 0,1 => (0, 1, 0, 1, 1)
3,1 -> 0,1 => (-1, -1, 0, -1, -1)
Standalone flags:
1,1.2 => (0, 1, 0, 0, 0)
1,1.3 => (-1, -1, 0, -2, -2)
1,1.4 => (0, 0, 0, 1, 0)
1,1.5 => (0, 0, 0, 0, 1)
0,1.4 => (0, 0, 0, 1, 0)
0,1.5 => (0, 0, 0, 0, 1)
2,1.2 => (1, 0, 0, 0, 0)
2,1.3 => (-1, -1, 0, -2, -2)
2,1.4 => (0, 0, 0, 1, 0)
2,1.5 => (0, 0, 0, 0, 1)
3,1.2 => (1, 0, 0, 0, 0)
3,1.3 => (0, 1, 0, 0, 0)
3,1.4 => (0, 0, 0, 1, 0)
3,1.5 => (0, 0, 0, 0, 1)

GW-Gauge correspondence

GW_CY5.gkm_3d_gauge — Function

gkm_3d_gauge(N::Int64, m::Int64)

Return the GKM graph of the CY3 space $X_{N,m}$ from [HS26, Section 3.1]. The returned space is Calabi-Yau but not equivariantly Calabi-Yau.

This function is used by gkm_5d_gauge to construct the GKM graph of $Z_{N,m} = X_{N,m}\times\mathbb{C}^2$.

Example

julia> G = gkm_3d_gauge(3, 5)
GKM graph with 6 nodes, valency 3 and axial function:
v2 -> v1 => (1, 1, 0)
v3 -> v2 => (1, 2, 0)
w1 -> v1 => (0, 0, -1)
w2 -> v2 => (0, 0, -1)
w2 -> w1 => (1, 1, 6)
w3 -> v3 => (0, 0, -1)
w3 -> w2 => (1, 2, 4)
Standalone flags:
v1.2 => (1, 0, 0)
v3.3 => (-1, -3, 0)
w1.2 => (1, 0, 8)
w3.3 => (-1, -3, -2)

julia> print_curve_classes(G)
v2 -> v1: (1, 0, 0), Chern number: 0
v3 -> v2: (0, 0, 1), Chern number: 0
w1 -> v1: (6, 1, 4), Chern number: 0
w2 -> v2: (0, 1, 4), Chern number: 0
w2 -> w1: (1, 0, 0), Chern number: 0
w3 -> v3: (0, 1, 0), Chern number: 0
w3 -> w2: (0, 0, 1), Chern number: 0

julia> first_chern_class(G)
(-t2 + t3)*e[1] + (-t2 + t3)*e[2] + (-t2 + t3)*e[3] + (-t2 + t3)*e[4] + (-t2 + t3)*e[5] + (-t2 + t3)*e[6]

GW_CY5.gkm_5d_gauge — Function

gkm_5d_gauge(N::Int64, m::Int64; equiCY::Bool=false)

Return the GKM graph of the CY5 space ${Z}_{N,m}$ from [HS26, Section 3.1]. If the optional argument equiCY is set to true then the substitution $\epsilon_2=\epsilon_3+\epsilon_4+\epsilon_5$ is applied to make the result equivariantly Calabi-Yau.

Examples

Let us first see the non-equivariantly Calabi-Yau example.

julia> G = gkm_5d_gauge(3, 5)
GKM graph with 6 nodes, valency 5 and axial function:
v2 -> v1 => (1, 1, 0, 0, 0)
v3 -> v2 => (1, 2, 0, 0, 0)
w1 -> v1 => (0, 0, -1, 0, 0)
w2 -> v2 => (0, 0, -1, 0, 0)
w2 -> w1 => (1, 1, 6, 0, 0)
w3 -> v3 => (0, 0, -1, 0, 0)
w3 -> w2 => (1, 2, 4, 0, 0)
Standalone flags:
v1.2 => (1, 0, 0, 0, 0)
v1.4 => (0, 0, 0, 1, 0)
v1.5 => (0, 0, 0, 0, 1)
v2.4 => (0, 0, 0, 1, 0)
v2.5 => (0, 0, 0, 0, 1)
v3.3 => (-1, -3, 0, 0, 0)
v3.4 => (0, 0, 0, 1, 0)
v3.5 => (0, 0, 0, 0, 1)
w1.2 => (1, 0, 8, 0, 0)
w1.4 => (0, 0, 0, 1, 0)
w1.5 => (0, 0, 0, 0, 1)
w2.4 => (0, 0, 0, 1, 0)
w2.5 => (0, 0, 0, 0, 1)
w3.3 => (-1, -3, -2, 0, 0)
w3.4 => (0, 0, 0, 1, 0)
w3.5 => (0, 0, 0, 0, 1)

julia> first_chern_class(G)
(-t2 + t3 + t4 + t5)*e[1] + (-t2 + t3 + t4 + t5)*e[2] + (-t2 + t3 + t4 + t5)*e[3] + (-t2 + t3 + t4 + t5)*e[4] + (-t2 + t3 + t4 + t5)*e[5] + (-t2 + t3 + t4 + t5)*e[6]

Second, let us see the equivariantly Calabi-Yau example.

julia> G = gkm_5d_gauge(3, 5; equiCY=true)
GKM graph with 6 nodes, valency 5 and axial function:
v2 -> v1 => (1, 0, 1, 1, 1)
v3 -> v2 => (1, 0, 2, 2, 2)
w1 -> v1 => (0, 0, -1, 0, 0)
w2 -> v2 => (0, 0, -1, 0, 0)
w2 -> w1 => (1, 0, 7, 1, 1)
w3 -> v3 => (0, 0, -1, 0, 0)
w3 -> w2 => (1, 0, 6, 2, 2)
Standalone flags:
v1.2 => (1, 0, 0, 0, 0)
v1.4 => (0, 0, 0, 1, 0)
v1.5 => (0, 0, 0, 0, 1)
v2.4 => (0, 0, 0, 1, 0)
v2.5 => (0, 0, 0, 0, 1)
v3.3 => (-1, 0, -3, -3, -3)
v3.4 => (0, 0, 0, 1, 0)
v3.5 => (0, 0, 0, 0, 1)
w1.2 => (1, 0, 8, 0, 0)
w1.4 => (0, 0, 0, 1, 0)
w1.5 => (0, 0, 0, 0, 1)
w2.4 => (0, 0, 0, 1, 0)
w2.5 => (0, 0, 0, 0, 1)
w3.3 => (-1, 0, -5, -3, -3)
w3.4 => (0, 0, 0, 1, 0)
w3.5 => (0, 0, 0, 0, 1)

julia> first_chern_class(G)
0

From 3-folds

GW_CY5.X_times_Ar — Function

X_times_Ar(X::AbstractGKM_graph, r::Int64; equiCY::Bool=false) -> GKMtools.AbstractGKM_graph

Return $X\times\mathcal{A}_r$.

Arguments

X: The GKM graph of a 3-fold.

r: The number of $T$-stable rational curves in $\mathcal{A}_r$.

equiCY: If this optional argument with default value false is set to true, the returned space will be linearized to be equivariantly Calabi-Yau.

Warning

For equiCY=true, this function requires X to be Calabi-Yau, although not necessarily equivariantly.

Example

Let us produce the example $\mathcal{A}_2\times\mathbb{C}\times\mathcal{A}_3$.

julia> A2_times_C = Ar_times_C1(2)
GKM graph with 3 nodes, valency 3 and axial function:
2 -> 1 => (1, -1, 0)
3 -> 2 => (1, -2, 0)
Standalone flags:
1.2 => (0, 0, 1)
1.3 => (1, 0, 0)
2.3 => (0, 0, 1)
3.2 => (0, 0, 1)
3.3 => (-1, 3, 0)

julia> A2_times_A3 = X_times_Ar(A2_times_C, 3)
GKM graph with 12 nodes, valency 5 and axial function:
2,1 -> 1,1 => (1, -1, 0, 0, 0)
3,1 -> 2,1 => (1, -2, 0, 0, 0)
1,2 -> 1,1 => (0, 0, 0, 1, -1)
2,2 -> 2,1 => (0, 0, 0, 1, -1)
2,2 -> 1,2 => (1, -1, 0, 0, 0)
3,2 -> 3,1 => (0, 0, 0, 1, -1)
3,2 -> 2,2 => (1, -2, 0, 0, 0)
1,3 -> 1,2 => (0, 0, 0, 1, -2)
2,3 -> 2,2 => (0, 0, 0, 1, -2)
2,3 -> 1,3 => (1, -1, 0, 0, 0)
3,3 -> 3,2 => (0, 0, 0, 1, -2)
3,3 -> 2,3 => (1, -2, 0, 0, 0)
1,4 -> 1,3 => (0, 0, 0, 1, -3)
2,4 -> 2,3 => (0, 0, 0, 1, -3)
2,4 -> 1,4 => (1, -1, 0, 0, 0)
3,4 -> 3,3 => (0, 0, 0, 1, -3)
3,4 -> 2,4 => (1, -2, 0, 0, 0)
Standalone flags:
1,1.2 => (0, 0, 1, 0, 0)
1,1.3 => (1, 0, 0, 0, 0)
1,1.5 => (0, 0, 0, 1, 0)
2,1.3 => (0, 0, 1, 0, 0)
2,1.5 => (0, 0, 0, 1, 0)
3,1.2 => (0, 0, 1, 0, 0)
3,1.3 => (-1, 3, 0, 0, 0)
3,1.5 => (0, 0, 0, 1, 0)
1,2.2 => (0, 0, 1, 0, 0)
1,2.3 => (1, 0, 0, 0, 0)
2,2.3 => (0, 0, 1, 0, 0)
3,2.2 => (0, 0, 1, 0, 0)
3,2.3 => (-1, 3, 0, 0, 0)
1,3.2 => (0, 0, 1, 0, 0)
1,3.3 => (1, 0, 0, 0, 0)
2,3.3 => (0, 0, 1, 0, 0)
3,3.2 => (0, 0, 1, 0, 0)
3,3.3 => (-1, 3, 0, 0, 0)
1,4.2 => (0, 0, 1, 0, 0)
1,4.3 => (1, 0, 0, 0, 0)
1,4.5 => (0, 0, 0, -1, 4)
2,4.3 => (0, 0, 1, 0, 0)
2,4.5 => (0, 0, 0, -1, 4)
3,4.2 => (0, 0, 1, 0, 0)
3,4.3 => (-1, 3, 0, 0, 0)
3,4.5 => (0, 0, 0, -1, 4)

From 4-folds

GW_CY5.CY5_from_CY4 — Function

CY5_from_CY4(CY4::AbstractGKM_graph; equiCY::Bool=false)

Return $X\times\mathbb{C}^1$.

Arguments

X: The GKM graph of a 4-fold.

equiCY: If this optional argument with default value false is set to true, the returned space will be linearized to be equivariantly Calabi-Yau.

Warning

For equiCY=true, this function requires X to be Calabi-Yau, although not necessarily equivariantly.

Example

Let us consider the example with $X=\text{Tot}(\mathcal{O}_{\mathbb{P}^2}(-1)\oplus \mathcal{O}_{\mathbb{P}^2}(-2))$.

julia> X = total_space(vector_bundle_O(2, [-1, -2]))
GKM graph with 3 nodes, valency 4 and axial function:
2 -> 1 => (-1, 1, 0, 0, 0)
3 -> 1 => (-1, 0, 1, 0, 0)
3 -> 2 => (0, -1, 1, 0, 0)
Standalone flags:
1.3 => (0, 0, 0, 1, 0)
1.4 => (0, 0, 0, 0, 1)
2.3 => (1, -1, 0, 1, 0)
2.4 => (2, -2, 0, 0, 1)
3.3 => (1, 0, -1, 1, 0)
3.4 => (2, 0, -2, 0, 1)

julia> CY5_from_CY4(X)
GKM graph with 3 nodes, valency 5 and axial function:
2,1 -> 1,1 => (-1, 1, 0, 0, 0, 0)
3,1 -> 1,1 => (-1, 0, 1, 0, 0, 0)
3,1 -> 2,1 => (0, -1, 1, 0, 0, 0)
Standalone flags:
1,1.3 => (0, 0, 0, 1, 0, 0)
1,1.4 => (0, 0, 0, 0, 1, 0)
1,1.5 => (0, 0, 0, 0, 0, 1)
2,1.3 => (1, -1, 0, 1, 0, 0)
2,1.4 => (2, -2, 0, 0, 1, 0)
2,1.5 => (0, 0, 0, 0, 0, 1)
3,1.3 => (1, 0, -1, 1, 0, 0)
3,1.4 => (2, 0, -2, 0, 1, 0)
3,1.5 => (0, 0, 0, 0, 0, 1)

julia> CY5_from_CY4(X; equiCY=true) # Let us also see the equivariantly CY substitution.
GKM graph with 3 nodes, valency 5 and axial function:
2,1 -> 1,1 => (-1, 1, 0, 0, 0, 0)
3,1 -> 1,1 => (-1, 0, 1, 0, 0, 0)
3,1 -> 2,1 => (0, -1, 1, 0, 0, 0)
Standalone flags:
1,1.3 => (0, 0, 0, 1, 0, 0)
1,1.4 => (0, 0, 0, 0, 1, 0)
1,1.5 => (-2, 1, 1, -1, -1, 0)
2,1.3 => (1, -1, 0, 1, 0, 0)
2,1.4 => (2, -2, 0, 0, 1, 0)
2,1.5 => (-2, 1, 1, -1, -1, 0)
3,1.3 => (1, 0, -1, 1, 0, 0)
3,1.4 => (2, 0, -2, 0, 1, 0)
3,1.5 => (-2, 1, 1, -1, -1, 0)

Further examples

GW_CY5.gkm_5d_P2_111 — Function

gkm_5d_P2_111()

Return the GKM graph of $\text{Tot}\mathcal{O}_{\mathbb{P}^2}(-1)^{\oplus 3}$ with the specific equivariantly CY linearization as in the example below.

Example

julia> G = gkm_5d_P2_111()
GKM graph with 3 nodes, valency 5 and axial function:
2 -> 1 => (2, -2, 0, 0, 0)
3 -> 1 => (2, 0, -2, 0, 0)
3 -> 2 => (0, 2, -2, 0, 0)
Standalone flags:
1.3 => (2, 0, 0, 2, 0)
1.4 => (2, 0, 0, 0, 2)
1.5 => (0, -2, -2, -2, -2)
2.3 => (0, 2, 0, 2, 0)
2.4 => (0, 2, 0, 0, 2)
2.5 => (-2, 0, -2, -2, -2)
3.3 => (0, 0, 2, 2, 0)
3.4 => (0, 0, 2, 0, 2)
3.5 => (-2, -2, 0, -2, -2)

julia> first_chern_class(G)
0

GW_CY5.gkm_5d_P3_04 — Function

gkm_5d_P3_04()

Return the GKM graph of $\text{Tot}(\mathcal{O}_{\mathbb{P}^3}(0)\oplus \mathcal{O}_{\mathbb{P}^3}(-4))$ with the specific equivariantly CY linearization as in the example below.

Example

julia> G = gkm_5d_P3_04()
GKM graph with 4 nodes, valency 5 and axial function:
2 -> 1 => (2, -2, 0, 0, 0)
3 -> 1 => (2, 0, -2, 0, 0)
3 -> 2 => (0, 2, -2, 0, 0)
4 -> 1 => (2, 0, 0, -2, 0)
4 -> 2 => (0, 2, 0, -2, 0)
4 -> 3 => (0, 0, 2, -2, 0)
Standalone flags:
1.4 => (0, 0, 0, 0, 2)
1.5 => (6, -2, -2, -2, -2)
2.4 => (0, 0, 0, 0, 2)
2.5 => (-2, 6, -2, -2, -2)
3.4 => (0, 0, 0, 0, 2)
3.5 => (-2, -2, 6, -2, -2)
4.4 => (0, 0, 0, 0, 2)
4.5 => (-2, -2, -2, 6, -2)

julia> first_chern_class(G)
0

GW_CY5.gkm_5d_P3_13 — Function

gkm_5d_P3_13()

Return the GKM graph of $\text{Tot}(\mathcal{O}_{\mathbb{P}^3}(-1)\oplus \mathcal{O}_{\mathbb{P}^3}(-3))$ with the specific equivariantly CY linearization as in the example below.

Example

julia> G = gkm_5d_P3_13()
GKM graph with 4 nodes, valency 5 and axial function:
2 -> 1 => (2, -2, 0, 0, 0)
3 -> 1 => (2, 0, -2, 0, 0)
3 -> 2 => (0, 2, -2, 0, 0)
4 -> 1 => (2, 0, 0, -2, 0)
4 -> 2 => (0, 2, 0, -2, 0)
4 -> 3 => (0, 0, 2, -2, 0)
Standalone flags:
1.4 => (2, 0, 0, 0, 2)
1.5 => (4, -2, -2, -2, -2)
2.4 => (0, 2, 0, 0, 2)
2.5 => (-2, 4, -2, -2, -2)
3.4 => (0, 0, 2, 0, 2)
3.5 => (-2, -2, 4, -2, -2)
4.4 => (0, 0, 0, 2, 2)
4.5 => (-2, -2, -2, 4, -2)

julia> first_chern_class(G)
0

GW_CY5.gkm_5d_P3_22 — Function

gkm_5d_P3_22()

Return the GKM graph of $\text{Tot}\mathcal{O}_{\mathbb{P}^3}(-2)^{\oplus 2}$ with the specific equivariantly CY linearization as in the example below.

Example

julia> G = gkm_5d_P3_22()
GKM graph with 4 nodes, valency 5 and axial function:
2 -> 1 => (2, -2, 0, 0, 0)
3 -> 1 => (2, 0, -2, 0, 0)
3 -> 2 => (0, 2, -2, 0, 0)
4 -> 1 => (2, 0, 0, -2, 0)
4 -> 2 => (0, 2, 0, -2, 0)
4 -> 3 => (0, 0, 2, -2, 0)
Standalone flags:
1.4 => (4, 0, 0, 0, 2)
1.5 => (2, -2, -2, -2, -2)
2.4 => (0, 4, 0, 0, 2)
2.5 => (-2, 2, -2, -2, -2)
3.4 => (0, 0, 4, 0, 2)
3.5 => (-2, -2, 2, -2, -2)
4.4 => (0, 0, 0, 4, 2)
4.5 => (-2, -2, -2, 2, -2)

julia> first_chern_class(G)
0